refactor(ios): split tokenizer and label tables out of MobileChatContentParser
Behavior-preserving move (MobileChatContentParser.swift 1,210 -> 824
lines): MobileChatProseTokenizer moves verbatim to
Core/MobileChatProseTokenizer.swift; MobileChatToolLabels and
MobileChatSkillLabels move verbatim to Core/MobileChatLabels.swift.
Completes the parser-file split started in #14244. pbxproj wired with
explicit entries.
